Add ftdetect for Amazon AWS CloudFormation templates
This isn't perfect. It assumes all *.template files are JSON. Also, for better or worse, it uses setfiletype instead of 'set filetype', so it will only set *.template to JSON if Vim hasn't already set the filetype for *.template files. So far Vim doesn't set a filetype for *.template files so we are ok.

Finally, someone really needs to write a syntax file for AWS CloudFormation templates. This ftdetect just sets the ft to JSON, and not a CloudFormation-specific version of JSON.